Moved DeviceExt into its own module.
