+++ src/gfx/lib.rs
@@ -24,8 +24,9 @@ extern crate libc;
 extern crate "gfx_device_gl" as device;
 
 // public re-exports
-pub use render::{DeviceExt, ProgramError, ShaderSource, Renderer, DrawError};
+pub use render::{Renderer, DrawError};
 pub use render::batch;
+pub use render::device_ext::{DeviceExt, ShaderSource, ProgramError};
 pub use render::mesh::{Attribute, Mesh, VertexFormat};
 pub use render::mesh::{Slice, ToSlice};
 pub use render::mesh::SliceKind;
