+++ src/render/lib.rs
@@ -25,12 +25,15 @@ use std::mem;
 use device::attrib;
 use device::attrib::IntSize;
 use device::draw::CommandBuffer;
-use device::shade::{ProgramInfo, UniformValue, Stage, CreateShaderError, ShaderModel};
+use device::shade::{ProgramInfo, UniformValue};
 use device::target::{Rect, ClearData, Mirror, Mask, Access, Target};
 use render::batch::Batch;
 use render::mesh::SliceKind;
 
-
+/// Batches
+pub mod batch;
+/// Device extensions
+pub mod device_ext;
 /// Meshes
 pub mod mesh;
 /// Shaders
@@ -39,16 +42,14 @@ pub mod shade;
 pub mod state;
 /// Render targets
 pub mod target;
-/// Batches
-pub mod batch;
 
 
 const TRACKED_ATTRIBUTES: usize = 8;
 type CachedAttribute = (device::RawBufferHandle, attrib::Format);
 type Instancing = (device::InstanceCount, device::VertexCount);
 
-/// The internal state of the renderer. This is used as a cache to eliminate
-/// redundant state changes.
+/// The internal state of the renderer.
+/// This is used as a cache to eliminate redundant state changes.
 struct RenderState {
     is_frame_buffer_set: bool,
     frame: target::Frame,
@@ -74,6 +75,7 @@ impl RenderState {
     }
 }
 
+/// Temporary parameter storage, used for shader activation.
 struct ParamStorage {
     uniforms: Vec<UniformValue>,
     blocks  : Vec<device::RawBufferHandle>,
@@ -81,6 +83,7 @@ struct ParamStorage {
 }
 
 impl ParamStorage{
+    /// Create an empty parameter storage.
     fn new() -> ParamStorage {
         ParamStorage {
             uniforms: Vec::new(),
@@ -457,128 +460,5 @@ impl<C: CommandBuffer> Renderer<C> {
                 self.command_buffer.call_draw_indexed(prim_type, IntSize::U32, start, end - start, base, instances);
             },
         }
-    }
-}
-
-
-/// Program linking error
-#[derive(Copy, Clone, PartialEq, Debug)]
-pub enum ProgramError {
-    /// Unable to compile the vertex shader
-    Vertex(CreateShaderError),
-    /// Unable to compile the fragment shader
-    Fragment(CreateShaderError),
-    /// Unable to link
-    Link(()),
-}
-
-/// A type storing shader source for different graphics APIs and versions.
-#[allow(missing_docs)]
-#[derive(Copy, Clone, PartialEq, Debug)]
-pub struct ShaderSource<'a> {
-    pub glsl_120: Option<&'a [u8]>,
-    pub glsl_130: Option<&'a [u8]>,
-    pub glsl_140: Option<&'a [u8]>,
-    pub glsl_150: Option<&'a [u8]>,
-    pub glsl_430: Option<&'a [u8]>,
-    // TODO: hlsl_sm_N...
-    pub targets: &'a [&'a str],
-}
-
-impl<'a> ShaderSource<'a> {
-    /// Pick one of the stored versions that is the highest supported by the device.
-    pub fn choose(&self, model: ShaderModel) -> Result<&'a [u8], ()> {
-        // following https://www.opengl.org/wiki/Detecting_the_Shader_Model
-        let version = model.to_number();
-        Ok(match *self {
-            ShaderSource { glsl_430: Some(s), .. } if version >= 50 => s,
-            ShaderSource { glsl_150: Some(s), .. } if version >= 40 => s,
-            ShaderSource { glsl_130: Some(s), .. } if version >= 30 => s,
-            ShaderSource { glsl_120: Some(s), .. } if version >= 20 => s,
-            _ => return Err(()),
-        })
-    }
-}
-
-
-/// Backend extension trait for convenience methods
-pub trait DeviceExt: device::Device {
-    /// Create a new renderer
-    fn create_renderer(&mut self) -> Renderer<<Self as device::Device>::CommandBuffer>;
-    /// Create a new mesh from the given vertex data.
-    /// Convenience function around `create_buffer` and `Mesh::from_format`.
-    fn create_mesh<T: mesh::VertexFormat + Copy>(&mut self, data: &[T]) -> mesh::Mesh;
-    /// Create a simple program given a vertex shader with a fragment one.
-    fn link_program(&mut self, vs_code: &[u8], fs_code: &[u8])
-                    -> Result<device::ProgramHandle, ProgramError>;
-    /// Create a simple program given `ShaderSource` versions of vertex and
-    /// fragment shaders, chooss the matching versions for the device.
-    fn link_program_source(&mut self, vs_src: ShaderSource, fs_src: ShaderSource)
-                           -> Result<device::ProgramHandle, ProgramError>;
-}
-
-impl<D: device::Device> DeviceExt for D {
-    fn create_renderer(&mut self) -> Renderer<<D as device::Device>::CommandBuffer> {
-        Renderer {
-            command_buffer: CommandBuffer::new(),
-            data_buffer: device::draw::DataBuffer::new(),
-            common_array_buffer: self.create_array_buffer(),
-            draw_frame_buffer: self.create_frame_buffer(),
-            read_frame_buffer: self.create_frame_buffer(),
-            default_frame_buffer: device::get_main_frame_buffer(),
-            render_state: RenderState::new(),
-            parameters: ParamStorage::new(),
-        }
-    }
-
-    fn create_mesh<T: mesh::VertexFormat + Copy>(&mut self, data: &[T]) -> mesh::Mesh {
-        let nv = data.len();
-        debug_assert!(nv < {
-            use std::num::Int;
-            let val: device::VertexCount = Int::max_value();
-            val as usize
-        });
-        let buf = self.create_buffer_static(data);
-        mesh::Mesh::from_format(buf, nv as device::VertexCount)
-    }
-
-    fn link_program(&mut self, vs_code: &[u8], fs_code: &[u8])
-                    -> Result<device::ProgramHandle, ProgramError> {
-        let vs = match self.create_shader(Stage::Vertex, vs_code) {
-            Ok(s) => s,
-            Err(e) => return Err(ProgramError::Vertex(e)),
-        };
-        let fs = match self.create_shader(Stage::Fragment, fs_code) {
-            Ok(s) => s,
-            Err(e) => return Err(ProgramError::Fragment(e)),
-        };
-
-        self.create_program(&[vs, fs], None)
-            .map_err(|e| ProgramError::Link(e))
-    }
-
-    fn link_program_source(&mut self, vs_src: ShaderSource, fs_src: ShaderSource)
-                           -> Result<device::ProgramHandle, ProgramError> {
-        let model = self.get_capabilities().shader_model;
-        let err_model = CreateShaderError::ModelNotSupported;
-
-        let vs = match vs_src.choose(model) {
-            Ok(code) => match self.create_shader(Stage::Vertex, code) {
-                Ok(s) => s,
-                Err(e) => return Err(ProgramError::Vertex(e)),
-            },
-            Err(_) => return Err(ProgramError::Vertex(err_model))
-        };
-
-        let fs = match fs_src.choose(model) {
-            Ok(code) => match self.create_shader(Stage::Fragment, code) {
-                Ok(s) => s,
-                Err(e) => return Err(ProgramError::Fragment(e)),
-            },
-            Err(_) => return Err(ProgramError::Fragment(err_model))
-        };
-
-        self.create_program(&[vs, fs], Some(fs_src.targets))
-            .map_err(|e| ProgramError::Link(e))
     }
 }
